+++ typings/index.d.ts
@@ -1,18 +1,1 @@
-/// <reference path="custom-typings.d.ts" />
-
-type RequestIdleCallbackHandle = any;
-type RequestIdleCallbackOptions = {
-  timeout: number;
-};
-type RequestIdleCallbackDeadline = {
-  readonly didTimeout: boolean;
-  timeRemaining: () => number;
-};
-
-interface Window {
-  requestIdleCallback: (
-    callback: (deadline: RequestIdleCallbackDeadline) => void,
-    opts?: RequestIdleCallbackOptions,
-  ) => RequestIdleCallbackHandle;
-  cancelIdleCallback: (handle: RequestIdleCallbackHandle) => void;
-}
+// / <reference path="custom-typings.d.ts" />
